+++ elementpath/protocols.py
@@ -10,80 +10,134 @@
 """
 Define type hints protocols for XPath related objects.
 """
-from typing import overload, Any, Dict, Iterator, Iterable, List, Literal, \
-    Optional, Protocol, Sized, Hashable, Union, TypeVar, runtime_checkable
+from typing import overload, Any, Dict, Iterator, Iterable, Optional, Sequence, ItemsView, \
+    Protocol, Sized, Hashable, Union, TypeVar, Mapping, Tuple, Set, MutableMapping
 
 _T = TypeVar("_T")
 
 
-@runtime_checkable
-class ElementProtocol(Iterable['ElementProtocol'], Sized, Hashable, Protocol):
+class LxmlAttribProtocol(Protocol):
+    """A minimal protocol for attribute of lxml Element objects."""
+    def get(self, *args: Any, **kwargs: Any) -> Optional[str]: ...
+
+    def items(self) -> Sequence[Tuple[Any, Any]]: ...
+
+    def __contains__(self, key: Any) -> bool: ...
+
+    def __getitem__(self, key: Any) -> Any: ...
+
+    def __iter__(self) -> Iterator[Any]: ...
+
+    def __len__(self) -> int: ...
+
+
+AttribType = Union[
+    MutableMapping[str, Any],
+    MutableMapping[Optional[str], Any],
+    LxmlAttribProtocol,
+    'XsdAttributeGroupProtocol'
+]
+
+
+class ElementProtocol(Sized, Hashable, Protocol):
+    """A protocol for generic ElementTree elements."""
+
+    def __iter__(self) -> Iterator['ElementProtocol']: ...
+
     def find(
-        self, path: str, namespaces: Optional[Dict[str, str]] = ...
+            self, path: str, namespaces: Optional[Dict[str, str]] = ...
     ) -> Optional['ElementProtocol']: ...
     def iter(self, tag: Optional[str] = ...) -> Iterator['ElementProtocol']: ...
+
     @overload
-    def get(self, key: str, default: None = ...) -> Optional[str]: ...
-    # noinspection PyOverloads
+    def get(self, key: str) -> Optional[str]: ...
+
     @overload
     def get(self, key: str, default: _T) -> Union[str, _T]: ...
-    tag: str
-    attrib: Dict[str, Any]
-    text: Optional[str]
-    tail: Optional[str]
+
+    @property
+    def tag(self) -> str: ...
+
+    @property
+    def text(self) -> Optional[str]: ...
+
+    @property
+    def tail(self) -> Optional[str]: ...
+
+    @property
+    def attrib(self) -> AttribType: ...
 
 
-@runtime_checkable
+class EtreeElementProtocol(ElementProtocol, Protocol):
+    """A protocol for xml.etree.ElementTree elements."""
+    def __iter__(self) -> Iterator['EtreeElementProtocol']: ...
+
+    def find(
+            self, path: str, namespaces: Optional[Dict[str, str]] = ...
+    ) -> Optional['EtreeElementProtocol']: ...
+    def iter(self, tag: Optional[str] = ...) -> Iterator['EtreeElementProtocol']: ...
+
+    @property
+    def attrib(self) -> Dict[str, str]: ...
+
+
 class LxmlElementProtocol(ElementProtocol, Protocol):
+    """A protocol for lxml.etree elements."""
+    def __iter__(self) -> Iterator['LxmlElementProtocol']: ...
+
+    def find(
+            self, path: str, namespaces: Optional[Dict[str, str]] = ...
+    ) -> Optional['LxmlElementProtocol']: ...
+    def iter(self, tag: Optional[str] = ...) -> Iterator['LxmlElementProtocol']: ...
+
     def getroottree(self) -> 'LxmlDocumentProtocol': ...
     def getnext(self) -> Optional['LxmlElementProtocol']: ...
     def getparent(self) -> Optional['LxmlElementProtocol']: ...
     def getprevious(self) -> Optional['LxmlElementProtocol']: ...
-    def itersiblings(self, tag: Optional[str] = ..., preceding: bool = False,
-                     *tags: str) -> Iterable['LxmlElementProtocol']: ...
-    nsmap: Dict[Optional[str], str]
+    def itersiblings(self, tag: Optional[str] = ..., *tags: str,
+                     preceding: bool = False) -> Iterable['LxmlElementProtocol']: ...
+
+    @property
+    def nsmap(self) -> Dict[Optional[str], str]: ...
+
+    @property
+    def attrib(self) -> LxmlAttribProtocol: ...
 
 
-@runtime_checkable
-class DocumentProtocol(Iterable[ElementProtocol], Hashable, Protocol):
+class DocumentProtocol(Hashable, Protocol):
     def getroot(self) -> Optional[ElementProtocol]: ...
-    def parse(self, source: Any, *args: Any, **kwargs: Any) -> 'DocumentProtocol': ...
+    def parse(self, source: Any, *args: Any, **kwargs: Any) -> ElementProtocol: ...
     def iter(self, tag: Optional[str] = ...) -> Iterator[ElementProtocol]: ...
 
 
-@runtime_checkable
-class LxmlDocumentProtocol(Iterable[LxmlElementProtocol], Hashable, Protocol):
+class LxmlDocumentProtocol(Hashable, Protocol):
     def getroot(self) -> Optional[LxmlElementProtocol]: ...
-    def parse(self, source: Any, *args: Any, **kwargs: Any) -> 'LxmlDocumentProtocol': ...
+    def parse(self, source: Any, *args: Any, **kwargs: Any) -> LxmlElementProtocol: ...
     def iter(self, tag: Optional[str] = ...) -> Iterator[LxmlElementProtocol]: ...
 
 
-@runtime_checkable
-class XsdValidatorProtocol(Protocol):
+class XsdValidatorProtocol(Hashable, Protocol):
     def is_matching(self, name: Optional[str],
                     default_namespace: Optional[str] = None) -> bool: ...
-    xsd_version: Literal['1.0', '1.1']
-    name: Optional[str]
-    maps: 'GlobalMapsProtocol'
+
+    @property
+    def name(self) -> Optional[str]: ...
+
+    @property
+    def xsd_version(self) -> str: ...
+
+    @property
+    def maps(self) -> 'GlobalMapsProtocol': ...
 
 
-@runtime_checkable
-class XsdSchemaProtocol(XsdValidatorProtocol, ElementProtocol, Protocol):
-    tag: Literal['{http://www.w3.org/2001/XMLSchema}schema']
-    attrib: Dict[str, 'XsdAttributeProtocol']
-    text: None
-
-
-XMLSchemaProtocol = XsdSchemaProtocol  # for backward compatibility
-
-
-@runtime_checkable
 class XsdComponentProtocol(XsdValidatorProtocol, Protocol):
-    parent: Optional['XsdComponentProtocol']
+
+    @property
+    def parent(self) -> Optional['XsdComponentProtocol']: ...
 
 
-@runtime_checkable
 class XsdTypeProtocol(XsdComponentProtocol, Protocol):
+
     def is_simple(self) -> bool:
         """Returns `True` if it's a simpleType instance, `False` if it's a complexType."""
         ...
@@ -150,38 +204,111 @@ class XsdTypeProtocol(XsdComponentProtocol, Protocol):
         """
         ...
 
-    root_type: 'XsdTypeProtocol'
-    """
-    The type at base of the definition of the XSD type. For a special type is the type
-    itself. For an atomic type is the primitive type. For a list is the primitive type
-    of the item. For a union is the base union type. For a complex type is xs:anyType.
-    """
+    @property
+    def root_type(self) -> 'XsdTypeProtocol':
+        """
+        The type at base of the definition of the XSD type. For a special type is the type
+        itself. For an atomic type is the primitive type. For a list is the primitive type
+        of the item. For a union is the base union type. For a complex type is xs:anyType.
+        """
+        ...
 
 
-@runtime_checkable
 class XsdAttributeProtocol(XsdComponentProtocol, Protocol):
-    type: Optional[XsdTypeProtocol]
-    ref: Optional['XsdAttributeProtocol']
+
+    @property
+    def type(self) -> Optional[XsdTypeProtocol]: ...
+
+    @property
+    def ref(self) -> Optional[Any]: ...
 
 
-@runtime_checkable
+XsdXPathNodeType = Union['XsdSchemaProtocol', 'XsdElementProtocol']
+
+
+class XsdAttributeGroupProtocol(XsdComponentProtocol, Protocol):
+
+    @overload
+    def get(self, key: Optional[str]) -> Optional[XsdAttributeProtocol]: ...
+
+    @overload
+    def get(self, key: Optional[str], default: _T) -> Union[XsdAttributeProtocol, _T]: ...
+
+    def items(self) -> ItemsView[Optional[str], XsdAttributeProtocol]: ...
+
+    def __contains__(self, key: Optional[str]) -> bool: ...
+
+    def __getitem__(self, key: Optional[str]) -> XsdAttributeProtocol: ...
+
+    def __iter__(self) -> Iterator[Optional[str]]: ...
+
+    def __len__(self) -> int: ...
+
+    def __hash__(self) -> int: ...
+
+    @property
+    def ref(self) -> Optional[Any]: ...
+
+
 class XsdElementProtocol(XsdComponentProtocol, ElementProtocol, Protocol):
-    type: Optional[XsdTypeProtocol]
-    ref: Optional['XsdElementProtocol']
-    attrib: Dict[str, XsdAttributeProtocol]
-    text: None
+
+    def __iter__(self) -> Iterator['XsdElementProtocol']: ...
+
+    def find(
+            self, path: str, namespaces: Optional[Dict[str, str]] = ...
+    ) -> Optional[XsdXPathNodeType]: ...
+    def iter(self, tag: Optional[str] = ...) -> Iterator['XsdElementProtocol']: ...
+
+    @property
+    def name(self) -> Optional[str]: ...
+
+    @property
+    def type(self) -> Optional[XsdTypeProtocol]: ...
+
+    @property
+    def ref(self) -> Optional[Any]: ...
+
+    @property
+    def attrib(self) -> XsdAttributeGroupProtocol: ...
+
+
+GT = TypeVar("GT")
+XsdGlobalValue = Union[GT, Tuple[ElementProtocol, Any]]
 
 
 class GlobalMapsProtocol(Protocol):
-    types: Dict[str, XsdTypeProtocol]
-    attributes: Dict[str, XsdAttributeProtocol]
-    elements: Dict[str, XsdElementProtocol]
-    substitution_groups: Dict[str, List[XsdElementProtocol]]
+
+    @property
+    def types(self) -> Mapping[str, XsdGlobalValue[XsdTypeProtocol]]: ...
+
+    @property
+    def attributes(self) -> Mapping[str, XsdGlobalValue[XsdAttributeProtocol]]: ...
+
+    @property
+    def elements(self) -> Mapping[str, XsdGlobalValue[XsdElementProtocol]]: ...
+
+    @property
+    def substitution_groups(self) -> Mapping[str, Set[Any]]: ...
 
 
-__all__ = ['ElementProtocol', 'LxmlElementProtocol',
-           'DocumentProtocol', 'LxmlDocumentProtocol',
-           'XsdValidatorProtocol', 'XsdSchemaProtocol',
-           'XsdComponentProtocol', 'XsdTypeProtocol',
-           'XsdElementProtocol', 'XsdAttributeProtocol',
-           'GlobalMapsProtocol', 'XMLSchemaProtocol']
+class XsdSchemaProtocol(XsdValidatorProtocol, ElementProtocol, Protocol):
+
+    def __iter__(self) -> Iterator[XsdXPathNodeType]: ...
+
+    def find(
+            self, path: str, namespaces: Optional[Dict[str, str]] = ...
+    ) -> Optional[XsdXPathNodeType]: ...
+    def iter(self, tag: Optional[str] = ...) -> Iterator[XsdXPathNodeType]: ...
+
+    @property
+    def tag(self) -> str: ...
+
+    @property
+    def attrib(self) -> MutableMapping[Optional[str], 'XsdAttributeProtocol']: ...
+
+
+__all__ = ['ElementProtocol', 'EtreeElementProtocol', 'LxmlAttribProtocol',
+           'LxmlElementProtocol', 'DocumentProtocol', 'LxmlDocumentProtocol',
+           'XsdValidatorProtocol', 'XsdComponentProtocol', 'XsdTypeProtocol',
+           'XsdAttributeProtocol', 'XsdAttributeGroupProtocol',
+           'XsdElementProtocol', 'GlobalMapsProtocol', 'XsdSchemaProtocol',]