+++ elementpath/xpath30/xpath30_parser.py
@@ -15,9 +15,10 @@ Refs:
   - https://www.w3.org/TR/xpath-functions-30/
 """
 from copy import deepcopy
-from typing import Any, Dict, Optional
+from typing import Any, ClassVar, Dict, Optional, Tuple
 
 from ..namespaces import XPATH_MATH_FUNCTIONS_NAMESPACE
+from ..datatypes import QName
 from ..xpath2 import XPath2Parser
 
 
@@ -37,7 +38,7 @@ class XPath30Parser(XPath2Parser):
     """
     version = '3.0'
 
-    DEFAULT_NAMESPACES = {
+    DEFAULT_NAMESPACES: ClassVar[Dict[str, str]] = {
         'math': XPATH_MATH_FUNCTIONS_NAMESPACE, **XPath2Parser.DEFAULT_NAMESPACES
     }
     PATH_STEP_SYMBOLS = {
@@ -52,7 +53,7 @@ class XPath30Parser(XPath2Parser):
         'schema-attribute', 'schema-element', 'switch', 'text', 'typeswitch',
     }
 
-    function_signatures = XPath2Parser.function_signatures.copy()
+    function_signatures: Dict[Tuple[QName, int], str] = XPath2Parser.function_signatures.copy()
 
     decimal_formats: DecimalFormatsType = {
         None: {
@@ -89,16 +90,16 @@ class XPath30Parser(XPath2Parser):
         if not defuse_xml:
             self.defuse_xml = defuse_xml
 
-    def __repr__(self) -> str:
+    def __str__(self) -> str:
         args = []
         if self.decimal_formats != self.__class__.decimal_formats:
             args.append(f'decimal_formats={self.decimal_formats!r}')
         if not self.defuse_xml:
             args.append('defuse_xml=False')
         if not args:
-            return super().__repr__()
+            return super().__str__()
 
-        repr_string = super().__repr__()[:-1]
+        repr_string = super().__str__()[:-1]
         if repr_string.endswith('('):
             return f"{repr_string}{', '.join(args)})"
         return f"{repr_string}, {', '.join(args)})"