+++ tests/test_xpath1_parser.py
@@ -72,11 +72,14 @@ class XPath1ParserTest(xpath_test_class.XPathTestCase):
 
     def test_string_representation(self):
         parser = self.parser.__class__()
-        self.assertEqual(repr(parser), f"{parser.__class__.__name__}()")
+        self.assertEqual(
+            repr(parser), f"<{parser.__class__.__name__} object at {hex(id(parser))}>"
+        )
+        self.assertEqual(str(parser), f"{parser.__class__.__name__}()")
 
         parser = self.parser.__class__(namespaces={'tst': 'http://xpath.test/ns'})
         self.assertEqual(
-            repr(parser), f"{parser.__class__.__name__}({{'tst': 'http://xpath.test/ns'}})"
+            str(parser), f"{parser.__class__.__name__}({{'tst': 'http://xpath.test/ns'}})"
         )
 
     def test_parser_pickling(self):
@@ -136,32 +139,38 @@ class XPath1ParserTest(xpath_test_class.XPathTestCase):
 
     def test_token_classes(self):
         # Literals
-        self.check_token('(string)', 'literal', "'hello' string",
-                         f"_StringLiteral({self.parser!r}, 'hello')", 'hello')
-        self.check_token('(integer)', 'literal', "1999 integer",
-                         f"_IntegerLiteral({self.parser!r}, 1999)", 1999)
-        self.check_token('(float)', 'literal', "3.1415 float",
-                         f"_FloatLiteral({self.parser!r}, 3.1415)", 3.1415)
-        self.check_token('(decimal)', 'literal', "217.35 decimal",
-                         f"_DecimalLiteral({self.parser!r}, 217.35)", 217.35)
-        self.check_token('(name)', 'literal', "'schema' name",
-                         f"_NameLiteral({self.parser!r}, 'schema')", 'schema')
+        self.check_token(
+            '(string)', 'literal', "'hello' string",
+            value='hello'
+        )
+        self.check_token(
+            '(integer)', 'literal', "1999 integer",
+            value=1999
+        )
+        self.check_token(
+            '(float)', 'literal', "3.1415 float",
+            value=3.1415
+        )
+        self.check_token(
+            '(decimal)', 'literal', "217.35 decimal",
+            value=217.35
+        )
+        self.check_token(
+            '(name)', 'literal', "'schema' name",
+            value='schema'
+        )
 
         # Variables
-        self.check_token('$', 'operator', "$ variable reference",
-                         f"_DollarSignOperator({self.parser!r})")
+        self.check_token('$', 'operator', "$ variable reference")
 
         # Axes
-        self.check_token('self', 'axis', "'self' axis",
-                         f"_SelfAxis({self.parser!r})")
-        self.check_token('child', 'axis', "'child' axis",
-                         f"_ChildAxis({self.parser!r})")
-        self.check_token('parent', 'axis', "'parent' axis",
-                         f"_ParentAxis({self.parser!r})")
-        self.check_token('ancestor', 'axis', "'ancestor' axis",
-                         f"_AncestorAxis({self.parser!r})")
-        self.check_token('preceding', 'axis', "'preceding' axis",
-                         expected_repr=f"_PrecedingAxis({self.parser!r})")
+        self.check_token('self', 'axis', "'self' axis")
+        self.check_token('child', 'axis', "'child' axis")
+        self.check_token('parent', 'axis', "'parent' axis")
+        self.check_token('ancestor', 'axis', "'ancestor' axis")
+        self.check_token(
+            'preceding', 'axis', "'preceding' axis"
+        )
         self.check_token('descendant-or-self', 'axis', "'descendant-or-self' axis")
         self.check_token('following-sibling', 'axis', "'following-sibling' axis")
         self.check_token('preceding-sibling', 'axis', "'preceding-sibling' axis")
@@ -176,22 +185,15 @@ class XPath1ParserTest(xpath_test_class.XPathTestCase):
         self.check_token(
             symbol='position',
             expected_label='function',
-            expected_str="'position' function",
-            expected_repr=f"_PositionFunction({self.parser!r})"
+            expected_str="'fn:position' function",
         )
 
         # Operators
-        self.check_token(
-            'and', 'operator', "'and' operator", f"_AndOperator({self.parser!r})"
-        )
+        self.check_token('and', 'operator', "'and' operator")
         if self.parser.version == '1.0':
-            self.check_token(
-                ',', 'symbol', "comma symbol", f"_CommaSymbol({self.parser!r})"
-            )
+            self.check_token(',', 'symbol', "comma symbol")
         else:
-            self.check_token(
-                ',', 'operator', "comma operator", f"_CommaOperator({self.parser!r})"
-            )
+            self.check_token(',', 'operator', "comma operator")
 
     def test_token_tree(self):
         self.check_tree('child::B1', '(child (B1))')
@@ -818,7 +820,7 @@ class XPath1ParserTest(xpath_test_class.XPathTestCase):
         self.check_value("boolean('   ')", True)
         self.check_value("boolean('')", False)
 
-        self.wrong_type("true(1)", 'XPST0017', "'true' function has no arguments")
+        self.wrong_type("true(1)", 'XPST0017', "'fn:true' function has no arguments")
         self.wrong_syntax("true(", 'unexpected end of source')
 
         if self.parser.version == '1.0':
@@ -1654,7 +1656,8 @@ class XPath1ParserTest(xpath_test_class.XPathTestCase):
 
     def test_get_function(self):
         func = self.parser.get_function('fn:true', 0)
-        self.assertEqual("'true' function", str(func))
+        self.assertEqual(str(func), "'fn:true' function")
+        self.assertTrue(repr(func).startswith('<XPathFunction fn:true#0 at 0x'))
         self.assertTrue(func())
 
         with self.assertRaises(NameError) as ctx:
@@ -1663,14 +1666,14 @@ class XPath1ParserTest(xpath_test_class.XPathTestCase):
 
         with self.assertRaises(TypeError) as ctx:
             self.parser.get_function('fn:true', 1)
-        self.assertIn('unknown function true#1', str(ctx.exception))
+        self.assertIn('unknown function fn:true#1', str(ctx.exception))
 
         func = self.parser.get_function('fn:false', 0)
-        self.assertEqual("'false' function", str(func))
+        self.assertTrue(repr(func).startswith('<XPathFunction fn:false#0 at 0x'))
         self.assertFalse(func())
 
         func = self.parser.get_function('concat', 2)
-        self.assertEqual("'concat' function", str(func))
+        self.assertTrue(repr(func).startswith('<XPathFunction fn:concat#2 at 0x'))
         self.assertEqual(func('foo', 1), 'foo1')
 
         with self.assertRaises(TypeError) as ctx:
@@ -1682,8 +1685,10 @@ class XPath1ParserTest(xpath_test_class.XPathTestCase):
         self.assertIn('too many arguments', str(ctx.exception))
 
         func = self.parser.get_function('concat', 4)
-        self.assertEqual("'concat' function", str(func))
+        self.assertTrue(repr(func).startswith('<XPathFunction fn:concat#4 at 0x'))
         self.assertEqual(func('foo', 1, ' bar', 2), 'foo1 bar2')
+
+        self.assertEqual(func.as_function()('foo', 1, ' bar', 2), 'foo1 bar2')
 
 
 @unittest.skipIf(lxml_etree is None, "The lxml library is not installed")