+++ tests/test_xpath30.py
@@ -98,7 +98,10 @@ class XPath30ParserTest(test_xpath2_parser.XPath2ParserTest):
         }
         self.assertDictEqual(parser.decimal_formats[None], expected)
         self.assertEqual(
-            repr(parser),
+            repr(parser), f"<{parser.__class__.__name__} object at {hex(id(parser))}>"
+        )
+        self.assertEqual(
+            str(parser),
             f'{parser.__class__.__name__}(decimal_formats={parser.decimal_formats})'
         )
 
@@ -111,17 +114,26 @@ class XPath30ParserTest(test_xpath2_parser.XPath2ParserTest):
     def test_defuse_xml_argument(self):
         parser = self.parser.__class__(defuse_xml=False)
         self.assertFalse(parser.defuse_xml)
-        self.assertEqual(repr(parser), f'{parser.__class__.__name__}(defuse_xml=False)')
+        self.assertEqual(
+            repr(parser), f'<{parser.__class__.__name__} object at {hex(id(parser))}>'
+        )
+        self.assertEqual(str(parser), f'{parser.__class__.__name__}(defuse_xml=False)')
 
         parser = self.parser.__class__({'tst': 'http://xpath.test/ns'}, defuse_xml=False)
         self.assertEqual(
-            repr(parser),
+            str(parser),
+            f"{parser.__class__.__name__}({{'tst': 'http://xpath.test/ns'}}, defuse_xml=False)"
+        )
+        self.assertEqual(
+            str(parser),
             f"{parser.__class__.__name__}({{'tst': 'http://xpath.test/ns'}}, defuse_xml=False)"
         )
 
     def test_function_match(self):
         token = self.parser.parse('math:pi()')
-        self.assertEqual(repr(token), f"_PrefixedReferenceToken({self.parser}, 'math:pi')")
+        self.assertEqual(
+            repr(token), f"<{token.__class__.__name__} object at {hex(id(token))}>"
+        )
         self.assertEqual(str(token), "'math:pi' function")
         self.assertEqual(token.source, 'math:pi()')
 
@@ -439,7 +451,9 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = 'math:sin(-math:pi() div 2)'
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token[1]), f"_MathSinFunction({self.parser})")
+        self.assertEqual(
+            repr(token[1]), f'<XPathFunction math:sin#1 at {hex(id(token[1]))}>'
+        )
         self.assertEqual(str(token[1]), "'math:sin' function")
 
     def test_cos_math_function(self):
@@ -458,7 +472,9 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = "math:cos(xs:double('INF'))"
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token[1]), f"_MathCosFunction({self.parser})")
+        self.assertEqual(
+            repr(token[1]), f'<XPathFunction math:cos#1 at {hex(id(token[1]))}>'
+        )
         self.assertEqual(str(token[1]), "'math:cos' function")
 
     def test_tan_math_function(self):
@@ -482,7 +498,9 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = 'math:tan(-0.0e0)'
         token = self.parser.parse(expression)
         self.assertEqual(token.source, 'math:tan(-0.0)')
-        self.assertEqual(repr(token[1]), f"_MathTanFunction({self.parser})")
+        self.assertEqual(
+            repr(token[1]), f"<XPathFunction math:tan#1 at {hex(id(token[1]))}>"
+        )
         self.assertEqual(str(token[1]), "'math:tan' function")
 
     def test_asin_math_function(self):
@@ -557,9 +575,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = 'fn:analyze-string("The cat sat on the mat.", "unmatchable")'
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression.replace('"', "'"))
-        self.assertEqual(repr(token[1]), f"_AnalyzeStringFunction({self.parser!r})")
-        self.assertEqual(str(token), "'fn:analyze-string' function")
-        self.assertEqual(str(token[1]), "'analyze-string' function")
+        self.assertEqual(
+            repr(token[1]), f'<XPathFunction fn:analyze-string at {hex(id(token[1]))}>'
+        )
+        self.assertEqual(str(token[1]), "'fn:analyze-string' function")
 
         context = XPathContext(root=self.etree.XML('<root/>'))
 
@@ -629,8 +648,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = 'has-children($elem/b1)'
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token), f"_HasChildrenFunction({self.parser!r})")
-        self.assertEqual(str(token), "'has-children' function")
+        self.assertEqual(
+            repr(token), f'<XPathFunction fn:has-children at {hex(id(token))}>'
+        )
+        self.assertEqual(str(token), "'fn:has-children' function")
 
     def test_innermost_function(self):
         with self.assertRaises(MissingContextError):
@@ -680,8 +701,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = 'innermost($nodes)'
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token), f"_InnermostFunction({self.parser!r})")
-        self.assertEqual(str(token), "'innermost' function")
+        self.assertEqual(
+            repr(token), f'<XPathFunction fn:innermost#1 at {hex(id(token))}>'
+        )
+        self.assertEqual(str(token), "'fn:innermost' function")
 
     def test_outermost_function(self):
         with self.assertRaises(MissingContextError):
@@ -743,8 +766,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = 'outermost($nodes)'
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token), f"_OutermostFunction({self.parser!r})")
-        self.assertEqual(str(token), "'outermost' function")
+        self.assertEqual(
+            repr(token), f'<XPathFunction fn:outermost#1 at {hex(id(token))}>'
+        )
+        self.assertEqual(str(token), "'fn:outermost' function")
 
     def test_parse_xml_function(self):
         with self.assertRaises(MissingContextError):
@@ -779,8 +804,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = "parse-xml('<alpha>abcd<alpha>')"
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token), f"_ParseXmlFunction({self.parser!r})")
-        self.assertEqual(str(token), "'parse-xml' function")
+        self.assertEqual(
+            repr(token), f'<XPathFunction fn:parse-xml#1 at {hex(id(token))}>'
+        )
+        self.assertEqual(str(token), "'fn:parse-xml' function")
 
     def test_parse_xml_fragment_function(self):
         root = self.etree.XML('<root/>')
@@ -851,8 +878,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = "parse-xml-fragment(' ')"
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token), f"_ParseXmlFragmentFunction({self.parser!r})")
-        self.assertEqual(str(token), "'parse-xml-fragment' function")
+        self.assertEqual(
+            repr(token), f'<XPathFunction fn:parse-xml-fragment#1 at {hex(id(token))}>'
+        )
+        self.assertEqual(str(token), "'fn:parse-xml-fragment' function")
 
     def test_serialize_function(self):
         root = self.etree.XML('<root/>')
@@ -979,7 +1008,9 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
             token.evaluate()
 
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token), f"_InlineFunction({self.parser!r})")
+        self.assertEqual(
+            repr(token), f'<_InlineFunction object at {hex(id(token))}>'
+        )
         self.assertEqual(str(token), "inline function")
 
         root = self.etree.XML('<root/>')
@@ -1008,8 +1039,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         token = self.parser.parse(expression)
         self.assertEqual(token.evaluate(), "bcd")
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token[0][1]), f"_FunctionLookupFunction({self.parser!r})")
-        self.assertEqual(str(token[0][1]), "'function-lookup' function")
+        self.assertEqual(
+            repr(token[0][1]), f'<XPathFunction fn:function-lookup#2 at {hex(id(token[0][1]))}>'
+        )
+        self.assertEqual(str(token[0][1]), "'fn:function-lookup' function")
 
         with self.xsd_version_parser('1.1'):
             token = self.parser.parse("(fn:function-lookup(xs:QName('xs:dateTimeStamp'), 1), "
@@ -1031,8 +1064,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = "fn:function-name(function($node) {count($node/*)})"
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token[1]), f"_FunctionNameFunction({self.parser!r})")
-        self.assertEqual(str(token[1]), "'function-name' function")
+        self.assertEqual(
+            repr(token[1]), f'<XPathFunction fn:function-name#1 at {hex(id(token[1]))}>'
+        )
+        self.assertEqual(str(token[1]), "'fn:function-name' function")
 
         # Context is not used if the argument is a function
         self.assertEqual(token.evaluate(), [])
@@ -1047,8 +1082,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = "fn:function-arity(function($node) {name($node)})"
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token[1]), f"_FunctionArityFunction({self.parser!r})")
-        self.assertEqual(str(token[1]), "'function-arity' function")
+        self.assertEqual(
+            repr(token[1]), f'<XPathFunction fn:function-arity#1 at {hex(id(token[1]))}>'
+        )
+        self.assertEqual(str(token[1]), "'fn:function-arity' function")
 
         # Context is not used if the argument is a function
         self.assertEqual(token.evaluate(), 1)
@@ -1060,8 +1097,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = 'fn:for-each(1 to 5, function($a) {$a * $a})'
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token[1]), f"_ForEachFunction({self.parser!r})")
-        self.assertEqual(str(token[1]), "'for-each' function")
+        self.assertEqual(
+            repr(token[1]), f'<XPathFunction fn:for-each#2 at {hex(id(token[1]))}>'
+        )
+        self.assertEqual(str(token[1]), "'fn:for-each' function")
 
         with self.assertRaises(MissingContextError):
             token.evaluate()
@@ -1080,8 +1119,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = 'fn:filter(1 to 10, function($a) {$a mod 2 = 0})'
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token[1]), f"_FilterFunction({self.parser!r})")
-        self.assertEqual(str(token[1]), "'filter' function")
+        self.assertEqual(
+            repr(token[1]), f'<XPathFunction fn:filter#2 at {hex(id(token[1]))}>'
+        )
+        self.assertEqual(str(token[1]), "'fn:filter' function")
 
         with self.assertRaises(MissingContextError):
             token.evaluate()
@@ -1094,8 +1135,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = 'fn:fold-left(1 to 5, 0, function($a, $b) {$a + $b})'
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token[1]), f"_FoldLeftFunction({self.parser!r})")
-        self.assertEqual(str(token[1]), "'fold-left' function")
+        self.assertEqual(
+            repr(token[1]), f'<XPathFunction fn:fold-left#3 at {hex(id(token[1]))}>'
+        )
+        self.assertEqual(str(token[1]), "'fn:fold-left' function")
 
         with self.assertRaises(MissingContextError):
             token.evaluate()
@@ -1131,8 +1174,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = 'fn:fold-right(1 to 5, 0, function($a, $b) {$a + $b})'
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression)
-        self.assertEqual(repr(token[1]), f"_FoldRightFunction({self.parser!r})")
-        self.assertEqual(str(token[1]), "'fold-right' function")
+        self.assertEqual(
+            repr(token[1]), f'<XPathFunction fn:fold-right#3 at {hex(id(token[1]))}>'
+        )
+        self.assertEqual(str(token[1]), "'fn:fold-right' function")
 
         with self.assertRaises(MissingContextError):
             token.evaluate()
@@ -1152,8 +1197,10 @@ class XPath30FunctionsTest(test_xpath2_functions.XPath2FunctionsTest):
         expression = 'fn:for-each-pair(("a", "b", "c"), ("x", "y", "z"), concat#2)'
         token = self.parser.parse(expression)
         self.assertEqual(token.source, expression.replace('"', "'"))
-        self.assertEqual(repr(token[1]), f"_ForEachPairFunction({self.parser!r})")
-        self.assertEqual(str(token[1]), "'for-each-pair' function")
+        self.assertEqual(
+            repr(token[1]), f'<XPathFunction fn:for-each-pair#3 at {hex(id(token[1]))}>'
+        )
+        self.assertEqual(str(token[1]), "'fn:for-each-pair' function")
         self.assertListEqual(token.evaluate(), ["ax", "by", "cz"])
 
         token = self.parser.parse('fn:for-each-pair(1 to 5, 1 to 5, function($a, $b){10*$a + $b})')