+++ tests/test_xpath31.py
@@ -109,11 +109,17 @@ class XPath31ParserTest(test_xpath30.XPath30ParserTest):
         self.assertEqual(token.symbol, 'map')
         self.assertEqual(token.label, 'map')
         self.assertEqual(token.source, f'map{map_value!r}'.replace(': ', ':'))
-        self.assertEqual(repr(token), f'XPathMap({self.parser!r}, None)')
+        self.assertEqual(
+            repr(token),
+            f"<XPathMap object (not evaluated constructor) at {hex(id(token))}>"
+        )
         self.assertEqual(str(token), 'not evaluated map constructor with 7 entries')
 
         self.assertDictEqual(token.evaluate()._map, map_value)
-        self.assertEqual(repr(token.evaluate()), f'XPathMap({self.parser!r}, {map_value!r})')
+
+        self.assertTrue(
+            repr(token.evaluate()).startswith('<XPathMap object at ')
+        )
         self.assertEqual(str(token.evaluate()), f'map{map_value!r}')
 
         token = self.parser.parse(f"{MAP_WEEKDAYS}('Mo')")
@@ -134,7 +140,9 @@ class XPath31ParserTest(test_xpath30.XPath30ParserTest):
         self.assertEqual(token.label, 'expression')
         self.assertTrue(token.source.startswith("map{'book':map{'title':'Data on the Web', "))
         self.assertTrue(token.source.endswith(", 'price':39.95}}('book')('title')"))
-        self.assertEqual(repr(token), f'_LeftParenthesisExpression({self.parser!r})')
+        self.assertEqual(
+            repr(token), f'<_LeftParenthesisExpression object at {hex(id(token))}>'
+        )
         self.assertEqual(str(token), "function call expression")
 
         token = self.parser.parse(f'{NESTED_MAP}("book")("author")')
@@ -188,11 +196,15 @@ class XPath31ParserTest(test_xpath30.XPath30ParserTest):
         self.assertEqual(token.symbol, 'array')
         self.assertEqual(token.label, 'array')
         self.assertEqual(token.source, 'array{1, 2, 5, 7}')
-        self.assertEqual(repr(token), f'XPathArray({self.parser!r}, None)')
+        self.assertEqual(
+            repr(token), f'<XPathArray object (not evaluated constructor) at {hex(id(token))}>'
+        )
         self.assertEqual(str(token), 'not evaluated curly array constructor with 4 items')
 
-        array = token.evaluate()
-        self.assertEqual(repr(array), f'XPathArray({self.parser!r}, [1, 2, 5, 7])')
+        array = token.evaluate()  # Create a new object ...
+        self.assertEqual(
+            repr(token), f'<XPathArray object (not evaluated constructor) at {hex(id(token))}>'
+        )
         self.assertEqual(str(array), '[1, 2, 5, 7]')
 
     def test_square_array_constructor(self):
@@ -202,31 +214,47 @@ class XPath31ParserTest(test_xpath30.XPath30ParserTest):
         self.assertEqual(token.symbol, '[')
         self.assertEqual(token.label, 'array')
         self.assertEqual(token.source, '[1, 2, 5, 7]')
-        self.assertEqual(repr(token), f'XPathArray({self.parser!r}, None)')
+        self.assertEqual(
+            repr(token),
+            f"<XPathArray object (not evaluated constructor) at {hex(id(token))}>"
+        )
         self.assertEqual(str(token), 'not evaluated square array constructor with 4 items')
 
         array = token.evaluate()
-        self.assertEqual(repr(array), f'XPathArray({self.parser!r}, [1, 2, 5, 7])')
+        self.assertEqual(
+            repr(array),
+            f"<{array.__class__.__name__} object at {hex(id(array))}>"
+        )
         self.assertEqual(str(array), '[1, 2, 5, 7]')
 
     def test_array_lookup(self):
         token = self.parser.parse('array { 1, 2, 5, 7 }(4)')
         self.assertEqual(token.evaluate(), 7)
         self.assertEqual(token.source, 'array{1, 2, 5, 7}(4)')
-        self.assertEqual(repr(token), f'_LeftParenthesisExpression({self.parser!r})')
+        self.assertEqual(repr(token), f'<_LeftParenthesisExpression object at {hex(id(token))}>')
+        self.assertEqual(
+            repr(token[0]),
+            f"<XPathArray object (not evaluated constructor) at {hex(id(token[0]))}>"
+        )
         self.assertEqual(str(token), "function call expression")
 
         token = self.parser.parse('[ 1, 2, 5, 7 ](4)')
         self.assertEqual(token.evaluate(), 7)
         self.assertEqual(token.source, '[1, 2, 5, 7](4)')
-        self.assertEqual(repr(token), f'_LeftParenthesisExpression({self.parser!r})')
+        self.assertEqual(repr(token), f'<_LeftParenthesisExpression object at {hex(id(token))}>')
+        self.assertEqual(
+            repr(token[0]),
+            f"<XPathArray object (not evaluated constructor) at {hex(id(token[0]))}>"
+        )
         self.assertEqual(str(token), "function call expression")
 
     def test_map_size_function(self):
         token = self.parser.parse('map:size(map{})')
         self.assertEqual(token.evaluate(), 0)
         self.assertEqual(str(token), "'map:size' function")
-        self.assertEqual(repr(token), f"_PrefixedReferenceToken({self.parser!r}, 'map:size')")
+        self.assertEqual(
+            repr(token), f'<_PrefixedReferenceToken object at {hex(id(token))}>'
+        )
         self.assertEqual(token.source, 'map:size(map{})')
 
         self.check_value('map:size(map{"true":1, "false":0})', 2)
@@ -235,7 +263,9 @@ class XPath31ParserTest(test_xpath30.XPath30ParserTest):
         token = self.parser.parse('map:keys(map{})')
         self.assertListEqual(token.evaluate(), [])
         self.assertEqual(str(token), "'map:keys' function")
-        self.assertEqual(repr(token), f"_PrefixedReferenceToken({self.parser!r}, 'map:keys')")
+        self.assertEqual(
+            repr(token), f'<_PrefixedReferenceToken object at {hex(id(token))}>'
+        )
         self.assertEqual(token.source, 'map:keys(map{})')
 
         self.check_value('map:keys(map{1:"yes", 2:"no"})', {1, 2})