+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 CHANGELOG
 *********
 
+`v4.4.0`_ (2024-03-11)
+======================
+* Improve stand-alone XPath functions builder (issue #70)
+* Update tokens and parsers __repr__
+* Fix static typing protocols to work with etree and XSD elements
+
 `v4.3.0`_ (2024-02-17)
 ======================
 * Change the purpose of the evaluation with a dynamic schema context
@@ -456,3 +462,4 @@ CHANGELOG
 .. _v4.2.0: https://github.com/sissaschool/elementpath/compare/v4.1.5...v4.2.0
 .. _v4.2.1: https://github.com/sissaschool/elementpath/compare/v4.2.0...v4.2.1
 .. _v4.3.0: https://github.com/sissaschool/elementpath/compare/v4.2.1...v4.3.0
+.. _v4.4.0: https://github.com/sissaschool/elementpath/compare/v4.3.0...v4.4.0